refactor/remove the MX standard endpoints
